\name{expected_sampling_time_for_genotype_list}
\alias{expected_sampling_time_for_genotype_list}
\title{expected_sampling_time_for_genotype_list}
\usage{
  expected_sampling_time_for_genotype_list(poset, lambdas,
    training_S_list, test_genotype_list, nr_sample_points)
}
\description{
  expected_sampling_time_for_genotype_list
}

